make it possible to connect ;send data; control the player ship and  shot lasers

+++ src/de/l_infotech/spaceinvader/game/SpaceObject.java
@@ -8,5 +8,5 @@ public interface SpaceObject {
 	
 	public boolean isAlive();
 	
-	public int[][] getGrafik();
+	public byte[][] getGrafik();
 }
